requirements.txt
PyYAML==5.4.1
tabulate==0.8.9
requests==2.25.1